Move cosign vars back to full-pipeline scope
Even though they're only needed in certain steps, the checking for
required env vars in init.sh expects them to be there.

In the long-term it would be better to only require the vars in the
steps where they're needed, but in the short term let's make this
change to get things working.

Ref: https://issues.redhat.com/browse/RHTAPBUGS-1279