support unset Rekor and TUF in GitHub Actions
Even when the secrets in GitHub are not set by user, the environment
variables still get created by GH Actions and are set to empty strings.
Check not just the existence of env vars, but also check that they are
not empty strings.
The values will then fall back to the defaults specified in env.sh