Display HTML BitBucket link in summary
For bitbucket, the Source/GitOps Repository template output links is
populated from the `remoteUrl` output from
the `publish-bitbucket-gitops` task which uses
the `publish:bitbucketCloud` action.

[This](https://github.com/backstage/backstage/blob/00292047ed0b3cc8659b8075c21281bd694ed911/plugins/scaffolder-backend-module-bitbucket-cloud/src/actions/bitbucketCloud.ts#L67-L89)
is how that plugin determines the value of `remoteUrl`.

The value is retrieved directly from the BitBucket's API. We can
replicate what the plugin is doing with curl and jq:

```
$ curl -s 'https://api.bitbucket.org/2.0/repositories/rhtap-test/bitbucket-tekton-quay-dotnet-14x' | jq -r '.links.clone[] | select(.name == "https") | .href'
https://bitbucket.org/rhtap-test/bitbucket-tekton-quay-dotnet-14x.git
```

Notice how this link is under the `.link.clone` section of the response.
This URL works just fine for cloning (as it is intended). It does not,
however, work for accessing the repo via a web browser.

For that, there's a different URL:

```
$ curl -s 'https://api.bitbucket.org/2.0/repositories/rhtap-test/bitbucket-tekton-quay-dotnet-14x' | jq -r '.links.html.href'
https://bitbucket.org/rhtap-test/bitbucket-tekton-quay-dotnet-14x
```

And that URL, of course, works just fine.

This commit changes the application creation template to use the
[repoContentsUrl](https://github.com/backstage/backstage/blob/00292047ed0b3cc8659b8075c21281bd694ed911/plugins/scaffolder-backend-module-bitbucket-cloud/src/actions/bitbucketCloud.ts#L93)
for this purpose.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>