Do not render gitops GH workflows
The GitHub Workflows used in the app and the gitops repos contain
variable placeholders that must not be resolved when creating the
component. Instead, these should be resolved by GitHub itself when the
workflow runs.

For example, the following should be copied as is without rendering the
variables:

```
IMAGE_REGISTRY: ${{ secrets.IMAGE_REGISTRY }}
```

This commit applies the same `copyWithoutRender` directive to the gitops
repo that is already applied to the app repo.

Ref: RHTAP-4047

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>