Remove import repo samples scripts
Prior to this change, the application sample repos were maintained
individually. Changes were performed directly at those repos. Once
changes on those were merged, someone would have to run the
`import-repo` script here to make those changes available to users.

Furhtermore, the sample repos being used are actually forks of other
sample repos that are no longer maintained and are planned to be
decommissioned.

In an effort to simplify maintenance, the sample application repos are
no longer used. Instead, the copies of the samples already in this
repository are now the source of truth.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>